# Fixup benchmark and speedup a little, fixes #141