--- a/Gemfile
+++ b/Gemfile
@@ -5,6 +5,7 @@ source 'https://rubygems.org/'
 gemspec
 
 group :benchmark do
+  gem 'benchmark-ips'
   gem 'kramdown'
   gem 'redcarpet'
 end