--- a/commonmarker.gemspec
+++ b/commonmarker.gemspec
@@ -23,6 +23,8 @@ Gem::Specification.new do |s|
   s.require_paths = %w[lib ext]
   s.required_ruby_version = ['>= 2.6', '< 4.0']
 
+  s.metadata['rubygems_mfa_required'] = 'true'
+
   s.rdoc_options += ['-x', 'ext/commonmarker/cmark/.*']
 
   s.add_development_dependency 'awesome_print'