--- a/test/benchmark.rb
+++ b/test/benchmark.rb
@@ -1,32 +1,39 @@
 # frozen_string_literal: true
 
+require 'benchmark/ips'
 require 'commonmarker'
-require 'github/markdown'
 require 'redcarpet'
 require 'kramdown'
 require 'benchmark'
 
-def dobench(name, &blk)
-  puts name
-  puts Benchmark.measure(&blk)
-end
+benchinput = File.read('test/benchinput.md').freeze
 
-benchinput = File.open('test/benchinput.md', 'r').read
+printf("input size = %<bytes>d bytes\n\n", { bytes: benchinput.bytesize })
 
-printf("input size = %<bytes>d bytes\n\n", benchinput.bytesize)
+Benchmark.ips do |x|
+  x.report('redcarpet') do
+    Redcarpet::Markdown.new(Redcarpet::Render::HTML, autolink: false, tables: false).render(benchinput)
+  end
 
-dobench('redcarpet') do
-  Redcarpet::Markdown.new(Redcarpet::Render::HTML, autolink: false, tables: false).render(benchinput)
-end
+  x.report('commonmarker with to_html') do
+    CommonMarker.render_html(benchinput)
+  end
 
-dobench('commonmarker with to_html') do
-  CommonMarker.render_html(benchinput)
-end
+  x.report('commonmarker with to_xml') do
+    CommonMarker.render_html(benchinput)
+  end
 
-dobench('commonmarker with ruby HtmlRenderer') do
-  CommonMarker::HtmlRenderer.new.render(CommonMarker.render_doc(benchinput))
-end
+  x.report('commonmarker with ruby HtmlRenderer') do
+    CommonMarker::HtmlRenderer.new.render(CommonMarker.render_doc(benchinput))
+  end
 
-dobench('kramdown') do
-  Kramdown::Document.new(benchinput).to_html(benchinput)
+  x.report('commonmarker with render_doc.to_html') do
+    CommonMarker.render_doc(benchinput, :DEFAULT, [:autolink]).to_html(:DEFAULT, [:autolink])
+  end
+
+  x.report('kramdown') do
+    Kramdown::Document.new(benchinput).to_html(benchinput)
+  end
+
+  x.compare!
 end