@ojab 2022-02-19
Fixed rubocop, hopefully you don't mind requiring mfa for rubygems.