@gjtorikian 2022-02-20
This all looks really great, thank you. As with any C change I want to take a deeper look at it (beyond the superficial one I've just given), but I don't see anything too outrageous. Thank you for your work. 😃